refact in the modal of shelfviewcontroller
