+++ ComicApp/Controller/PopUpModalView.swift
@@ -8,25 +8,3 @@
 
 import UIKit
 
-extension ShelfViewController: PopUpModalDelegate {
-    
-    func popUpModal(image: UIImage, comic: ComicCD) {
-        self.selectedComic = comic
-        prepareForTransition()
-        setBlurEffectView()
-        setImageForModal(fromImage: image)
-        setLableForTitleInModal(fromText: comic.title!)
-        let progressNumber = comic.progressNumber
-        let finishNumber = comic.finishNumber
-        if comic.status != "Quero Ler" {
-            setStatusForModal(status: "\(progressNumber)/\(finishNumber)", progress: (Float(progressNumber)/Float(finishNumber)), comicStatus: comic.status!)
-        }
-    }
-    
-     func prepareForTransition() {
-         if let comic = self.selectedComic {
-            // self.selectedIndex = listOfComics.firstIndex(of: comic)
-         }
-     }
-    
-}
